Rename base class to OAuthProvider from SocialMedia
As pointed by @xurizaemon, OAuthProvider makes more sense according to what the class does. SocialMedia can rather be confusing.
