+++ CRM/Civisocial/Backend/OAuthProvider/Twitter.php
@@ -1,8 +1,8 @@
 <?php
-require_once 'CRM/Civisocial/Backend/SocialMedia.php';
-require_once 'CRM/Civisocial/Backend/SocialMedia/OAuth/OAuth.php';
+require_once 'CRM/Civisocial/Backend/OAuthProvider.php';
+require_once 'CRM/Civisocial/Backend/OAuthProvider/OAuth/OAuth.php';
 
-class CRM_Civisocial_Backend_SocialMedia_Twitter extends CRM_Civisocial_Backend_SocialMedia {
+class CRM_Civisocial_Backend_OAuthProvider_Twitter extends CRM_Civisocial_Backend_OAuthProvider {
 
 	/**
 	 * Short name (alias) for OAuth provider
